+++ man/roxygen/meta.R
@@ -5,10 +5,10 @@ list( # nolint
     postprocess = "Functions used for postprocessing of model fits", # nolint
     plot = "Plotting functions",
     forecast = "Functions used for forecasting across models, dates, and scenarios", # nolint
-    generatedata = "Functions to generate simulated data.",
-    scenarios = "Functions to define and create data scenarios.",
-    modelvalidation = "Functions to explore and validate models.", # nolint
+    generatedata = "Functions to generate simulated data",
+    scenarios = "Functions to define and create data scenarios",
+    modelvalidation = "Functions to explore and validate models", # nolint
     data = "Package data sets",
-    check = "Functions used for checking inputs."
+    check = "Functions used for checking inputs"
   )
 )